refactor(flterm): simplify terminal frame construction
