+++ packages/flterm/test/rendering/frame_builder_test.dart
@@ -44,6 +44,11 @@ void main() {
       return List.generate(sprites.count, (index) => transforms[index * 4 + 2]);
     }
 
+    List<double> yPositions(AtlasSprites sprites) {
+      final transforms = sprites.sealedTransforms;
+      return List.generate(sprites.count, (index) => transforms[index * 4 + 3]);
+    }
+
     String entryRectKey(AtlasEntry entry) {
       return [
         entry.srcLeft,
@@ -240,6 +245,17 @@ void main() {
         hasLength(3),
       );
       expect(atlas.cacheSize, initialCacheSize);
+    });
+
+    test('sync separates operator runs at a foreground color boundary', () {
+      writeUtf8(terminal, '=>\x1b[31m!=');
+
+      builder.sync(terminal, terminalDirty: true);
+
+      expect(
+        sprites.shaped.rows.where((row) => row.isNotEmpty).single,
+        hasLength(2),
+      );
     });
 
     test('sync chunks long operator runs without adding atlas entries', () {
@@ -518,6 +534,47 @@ void main() {
       expect(sprites.regular.sealedColors, contains(0xFF00FF00.toSigned(32)));
     });
 
+    test('sync restores rows when search matches are cleared', () {
+      state.updateTheme(
+        TerminalTheme.dark().copyWith(
+          search: const SearchTheme(
+            match: SelectionTheme(
+              foreground: DynamicColor.fixed(Color(0xFFFF0000)),
+            ),
+          ),
+        ),
+      );
+      writeUtf8(terminal, 'AB\r\nCD');
+      builder.sync(terminal, terminalDirty: true);
+      final normalColors = sprites.regular.sealedColors.toList();
+      final match = Selection.fromRefs(
+        start: GridRef.at(terminal, const Position(row: 0, col: 0)),
+        end: GridRef.at(terminal, const Position(row: 0, col: 1)),
+      );
+
+      builder.sync(
+        terminal,
+        terminalDirty: false,
+        searchDirty: true,
+        searchMatches: [match],
+      );
+      builder.sync(terminal, terminalDirty: false, searchDirty: true);
+
+      expect(sprites.regular.sealedColors, normalColors);
+    });
+
+    test('sync resets hidden style state at a default row', () {
+      final frame = createFrame(cols: 1, rows: 2);
+      writeUtf8(frame.terminal, '\x1b[8;31mA\x1b[0m\r\nB');
+
+      frame.builder.sync(frame.terminal, terminalDirty: true);
+
+      expect(frame.sprites.regular.sealedColors, [
+        frame.state.terminalForegroundArgb.toSigned(32),
+      ]);
+      expect(yPositions(frame.sprites.regular), [16.0]);
+    });
+
     test('sync resolves block cursor glyph from cached frame state', () {
       writeUtf8(terminal, 'A\x1b[1;1H');
       builder.sync(terminal, terminalDirty: true);
@@ -722,6 +779,19 @@ void main() {
       expect(xPositions(sprites.regular), [0.0, 8.0, 16.0, 24.0, 32.0, 40.0]);
     });
 
+    test('sync moves unchanged preedit text with the cursor', () {
+      writeUtf8(terminal, 'A\r\nB\x1b[1;1H');
+      builder.sync(terminal, terminalDirty: true, preeditText: '日');
+      writeUtf8(terminal, '\x1b[2;1H');
+
+      builder.sync(terminal, terminalDirty: true, preeditText: '日');
+
+      expect(xPositions(sprites.regular), [0.0]);
+      expect(yPositions(sprites.regular), [0.0]);
+      expect(xPositions(sprites.wide), [0.0]);
+      expect(yPositions(sprites.wide), [16.0]);
+    });
+
     group('incremental invalidation', () {
       test('blink changes retain rows without blinking cells', () {
         writeUtf8(terminal, '\x1b[5m=>\x1b[0m\r\n!=');
@@ -733,6 +803,22 @@ void main() {
         builder.sync(terminal, terminalDirty: false);
 
         expect(sprites.shaped.rows[1].single, same(retainedRun));
+      });
+
+      test('blink changes hide and show blinking rows', () {
+        writeUtf8(terminal, '\x1b[5mA\x1b[0m\r\nB');
+        builder.sync(terminal, terminalDirty: true);
+
+        state.blinkVisible = false;
+        builder.markBlinkRowsDirty();
+        builder.sync(terminal, terminalDirty: false);
+        expect(sprites.regular.count, 1);
+
+        state.blinkVisible = true;
+        builder.markBlinkRowsDirty();
+        builder.sync(terminal, terminalDirty: false);
+
+        expect(sprites.regular.count, 2);
       });
 
       test('search changes retain rows outside matches', () {
